refactor: export module
